도메인 모델 정의 /Channel 클래스
