+++ src/main/java/com/sprint/mission/discodeit/entity/Channel.java
@@ -1,0 +1,19 @@
+package com.sprint.mission.discodeit.entity;
+
+public class Channel extends BaseEntity {
+    private String name;
+
+    public Channel(String name) {
+        super();
+        this.name = name;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void updateName(String name) {
+        this.name = name;
+        update();
+    }
+}
