+++ src/main/java/com/sprint/mission/discodeit/entity/User.java
@@ -1,19 +1,19 @@
 package com.sprint.mission.discodeit.entity;
 //사용자 정보를 관리
 public class User extends BaseEntity {
-    private String name;
+    private String username;
 
-    public User(String name) {
+    public User(String username) {
         super(); //부모 클래스 초기화
-        this.name = name;
+        this.username = username;
     }
 
-    public String getName() {
-        return name;
+    public String getUserName() {
+        return username;
     }
 
-    public void updateName(String name) {
-        this.name = name;
+    public void updateUserName(String username) {
+        this.username = username;
         update();
     }
